+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### New
 
+- [#1948](https://gitlab.com/meltano/meltano/-/issues/1948) Show Intercom button in bottom right on MeltanoData.com instances
+
 ### Changes
 
 ### Fixes