+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### New
 
 - [#1948](https://gitlab.com/meltano/meltano/-/issues/1948) Show Intercom button in bottom right on MeltanoData.com instances
+- [#1930](https://gitlab.com/meltano/meltano/-/issues/1930) Add button to remove report from dashboard when editing dashboard
 
 ### Changes