+++ CHANGELOG.md
@@ -15,6 +15,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### Changes
 
+- [!1607](https://gitlab.com/meltano/meltano/-/merge_requests/1607) Move date range picker into results area of report builder interface
+
 ### Fixes
 
 ### Breaks