+++ CHANGELOG.md
@@ -12,8 +12,11 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 - [#1948](https://gitlab.com/meltano/meltano/-/issues/1948) Show Intercom button in bottom right on MeltanoData.com instances
 - [#1930](https://gitlab.com/meltano/meltano/-/issues/1930) Add button to remove report from dashboard when editing dashboard
+- [#1845](https://gitlab.com/meltano/meltano/-/issues/1845) Add button to delete report to report builder interface
 
 ### Changes
+
+- [!1607](https://gitlab.com/meltano/meltano/-/merge_requests/1607) Move date range picker into results area of report builder interface
 
 ### Fixes