+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 - [#1930](https://gitlab.com/meltano/meltano/-/issues/1930) Add button to remove report from dashboard when editing dashboard
 - [#1845](https://gitlab.com/meltano/meltano/-/issues/1845) Add button to delete report to report builder interface
 - [#1849](https://gitlab.com/meltano/meltano/-/issues/1849) Add button to rename report to report builder interface
+- [#1951](https://gitlab.com/meltano/meltano/-/issues/1951) Add button to edit dashboard name and description to dashboard page
 
 ### Changes