+++ CHANGELOG.md
@@ -10,6 +10,18 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### New
 
+### Changes
+
+### Fixes
+
+### Breaks
+
+
+## 1.29.1 - (2020-04-16)
+---
+
+### New
+
 - [#1948](https://gitlab.com/meltano/meltano/-/issues/1948) Show Intercom button in bottom right on MeltanoData.com instances
 - [#1930](https://gitlab.com/meltano/meltano/-/issues/1930) Add button to remove report from dashboard when editing dashboard
 - [#1845](https://gitlab.com/meltano/meltano/-/issues/1845) Add button to delete report to report builder interface
@@ -20,10 +32,6 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 - [!1607](https://gitlab.com/meltano/meltano/-/merge_requests/1607) Move date range picker into results area of report builder interface
 - [!1608](https://gitlab.com/meltano/meltano/-/merge_requests/1608) Make report title more prominent in report builder
-
-### Fixes
-
-### Breaks
 
 
 ## 1.29.0 - (2020-04-13)